Removed the AptController and renamed aptLayout to demoModel
