+++ View/demoModel.py
@@ -1,14 +1,12 @@
 import tkinter as tk
 from PIL import Image, ImageTk, ImageEnhance
 from model.observer import Observer
-import controller.AptController as AptController
 import model.AptModel as AptModel
 import os
 
-class AptLayout(Observer):
-    def __init__(self, parent, controller : AptController, model : AptModel):
+class demoModel(Observer):
+    def __init__(self, parent, model : AptModel):
         super().__init__()
-        self.controller = controller
         self.model = model
         
         # Image directory and new height
@@ -61,7 +59,7 @@ class AptLayout(Observer):
 
     # Method to toggle room states
     def toggle_rooms_state(self, room_name, action):
-        self.controller.model.toggle_rooms_state(room_name, action)
+        self.model.toggle_rooms_state(room_name, action)
         self.update_layout()
 
     # Method to update the layout
@@ -85,6 +83,6 @@ class AptLayout(Observer):
 
     # Method to update the observer
     def notify(self):
-        self.controller.notify()
+        self.model.notify()
         self.update_layout()
         self.update_display()
